my-app/requirements-test.txt
pytest==7.4.2
pytest-cov==4.1.0
flake8==6.1.0
mypy==1.5.1
black==23.7.0
isort==5.12.0
pytest-flask==1.2.0
requests==2.32.5
